Improving one note and correcting another.
The first note change is because it seemed clunky* as I was reading it, so I'd like to propose those changes. 

The second note change is directly from customer feedback and to correct the misinterpretation that the panel object viewer can delete all objects within a bucket; it can't. Also felt like moving that notice up to the top of that section was important. 

* At least, that's the best word I can think of. But I'm not a professional writer and if I think about it too much my eye starts twitching lol.